arcdps::evtc

Enum EventKind

pub enum EventKind {

    EnterCombat(EnterCombatEvent),
    ExitCombat(AgentStatusEvent),
    ChangeUp(AgentStatusEvent),
    ChangeDead(AgentStatusEvent),
    ChangeDown(AgentStatusEvent),
    Spawn(AgentStatusEvent),
    Despawn(AgentStatusEvent),
    HealthUpdate(HealthUpdateEvent),
    SquadCombatStart(LogEvent),
    SquadCombatEnd(LogEvent),
    WeaponSwap(WeaponSwapEvent),
    MaxHealthUpdate(MaxHealthEvent),
    PointOfView(AgentStatusEvent),
    Language {
        time: u64,
        language: Result<Language, u64>,
    },
    GWBuild {
        time: u64,
        build: u64,
    },
    ShardId {
        time: u64,
        shard: u64,
    },
    Reward(RewardEvent),
    BuffInitial(BuffInitialEvent),
    Position(PositionEvent),
    Velocity(PositionEvent),
    Facing(PositionEvent),
    TeamChange(TeamChangeEvent),
    AttackTarget(AttackTargetEvent),
    Targetable(TargetableEvent),
    MapId {
        time: u64,
        map: u64,
    },
    StackActive(StackActiveEvent),
    StackReset(StackResetEvent),
    Guild(GuildEvent),
    BuffInfo(BuffInfo),
    BuffFormula(BuffFormula),
    SkillInfo(SkillInfo),
    SkillTiming(SkillTiming),
    BreakbarState(BreakbarStateEvent),
    BreakbarPercent(BreakbarPercentEvent),
    Integrity(ErrorEvent),
    AgentMarker(AgentMarkerEvent),
    BarrierUpdate(BarrierUpdateEvent),
    StatReset {
        time: u64,
        target: u64,
    },
    Extension {
        sig: u32,
        event: Event,
    },
    ApiDelayed {
        event: Box<EventKind>,
    },
    InstanceStart {
        time: u64,
        start: u64,
    },
    RateHealth {
        time: u64,
        rate: u64,
    },
    Last90BeforeDown(DownContributionEvent),
    Effect45(Effect45),
    IdToGUID(ContentInfo),
    LogNPCUpdate(LogEvent),
    ExtensionCombat {
        sig: u32,
        event: Event,
    },
    FractalScale {
        time: u64,
        scale: u64,
    },
    Effect51(Effect51),
    Ruleset(Ruleset),
    Activation(ActivationEvent),
    BuffRemove(BuffRemoveEvent),
    BuffApply(BuffApplyEvent),
    BuffDamage(BuffDamageEvent),
    Strike(StrikeEvent),
    SquadMarker(SquadMarkerEvent),
    ArcBuild(ArcBuildEvent),
    Glider(GliderEvent),
    Stunbreak(StunbreakEvent),
    MissileCreate(MissileCreate),
    MissileLaunch(MissileLaunch),
    MissileRemove(MissileRemove),
    EffectGroundCreate(GroundEffect),
    EffectGroundRemove(GroundEffectRemove),
    EffectAgentCreate(AgentEffect),
    EffectAgentRemove(AgentEffectRemove),
    Unknown(Event),

}

Possible Event kinds.

Variants

EnterCombat(EnterCombatEvent)

Agent entered combat.

ExitCombat(AgentStatusEvent)

Agent left combat.

ChangeUp(AgentStatusEvent)

Agent is now alive.

ChangeDead(AgentStatusEvent)

Agent is now dead.

ChangeDown(AgentStatusEvent)

Agent is now downed.

Spawn(AgentStatusEvent)

Agent is now in game tracking range.

Despawn(AgentStatusEvent)

Agent is no longer being tracked or out of game tracking range.

HealthUpdate(HealthUpdateEvent)

Agent health change.

SquadCombatStart(LogEvent)

Log started.

SquadCombatEnd(LogEvent)

Log ended.

WeaponSwap(WeaponSwapEvent)

Agent swapped weapon set.

MaxHealthUpdate(MaxHealthEvent)

Agent maximum health change.

PointOfView(AgentStatusEvent)

Agent is “recording” player.

Language

Game text language.

Fields

time: u64

language: Result<Language, u64>

GWBuild

Game build.

Fields

time: u64

build: u64

ShardId

Sever shard id.

Fields

time: u64

shard: u64

Reward(RewardEvent)

Agent got a reward chest.

BuffInitial(BuffInitialEvent)

Appears once per buff per agent on logging start.

Position(PositionEvent)

Agent position change.

Velocity(PositionEvent)

Agent velocity change.

Facing(PositionEvent)

Agent facing change.

TeamChange(TeamChangeEvent)

Agent team change.

AttackTarget(AttackTargetEvent)

Agent is now an attack target.

Targetable(TargetableEvent)

Agent targetability change.

MapId

Map id.

Fields

time: u64

map: u64

StackActive(StackActiveEvent)

Agent with active buff.

StackReset(StackResetEvent)

Agent with reset buff.

Guild(GuildEvent)

Agent is in guild.

BuffInfo(BuffInfo)

Buff information.

BuffFormula(BuffFormula)

Buff formula.

SkillInfo(SkillInfo)

Skill information.

SkillTiming(SkillTiming)

Skill action.

BreakbarState(BreakbarStateEvent)

Agent breakbar state change.

BreakbarPercent(BreakbarPercentEvent)

Breakbar percentage.

Integrity(ErrorEvent)

Error.

AgentMarker(AgentMarkerEvent)

Agent has marker.

BarrierUpdate(BarrierUpdateEvent)

Agent barrier change.

StatReset

Arc UI stats reset.

Fields

time: u64

target: u64

Extension

A custom event created by an extension (addon/plugin).

Fields

sig: u32

event: Event

ApiDelayed

Delayed combat event.

Fields

event: Box<EventKind>

InstanceStart

Instance started.

Fields

time: u64

start: u64

RateHealth

Tick rate.

Fields

time: u64

rate: u64

Last90BeforeDown(DownContributionEvent)

Last 90% before down for downs contribution.

Effect45(Effect45)

Effect created or ended.

IdToGUID(ContentInfo)

Content id to GUID.

This maps a volatile content id to a stable GUID.

LogNPCUpdate(LogEvent)

Log NPC changed.

ExtensionCombat

A custom combat event created by an extension (addon/plugin).

Fields

sig: u32

event: Event

FractalScale

Fractal scale.

Fields

time: u64

scale: u64

Effect51(Effect51)

Effect created or ended.

Ruleset(Ruleset)

Combat ruleset.

Activation(ActivationEvent)

Activation (cast) event.

BuffRemove(BuffRemoveEvent)

Buff removed.

BuffApply(BuffApplyEvent)

Buff applied.

BuffDamage(BuffDamageEvent)

Buff damage.

Strike(StrikeEvent)

Direct (strike) damage.

SquadMarker(SquadMarkerEvent)

Squad marker placed or removed.

ArcBuild(ArcBuildEvent)

ArcDPS build information.

Glider(GliderEvent)

Agent gliding state changed.

Stunbreak(StunbreakEvent)

Effect

MissileCreate(MissileCreate)

Missile created.

MissileLaunch(MissileLaunch)

Missile launched or relaunched.

MissileRemove(MissileRemove)

Missile removed or destroyed.

EffectGroundCreate(GroundEffect)

Ground effect created.

EffectGroundRemove(GroundEffectRemove)

Ground effect removed.

EffectAgentCreate(AgentEffect)

Effect around Agent created.

EffectAgentRemove(AgentEffectRemove)

Effect around Agent removed.

Unknown(Event)

Unknown event.