Enum arcdps::evtc::event::EventKind

pub enum EventKind {

    EnterCombat(EnterCombatEvent),
    ExitCombat(AgentStatusEvent),
    ChangeUp(AgentStatusEvent),
    ChangeDead(AgentStatusEvent),
    ChangeDown(AgentStatusEvent),
    Spawn(AgentStatusEvent),
    Despawn(AgentStatusEvent),
    HealthUpdate(HealthUpdateEvent),
    LogStart(LogEvent),
    LogEnd(LogEvent),
    WeaponSwap(WeaponSwapEvent),
    MaxHealthUpdate(MaxHealthEvent),
    PointOfView(AgentStatusEvent),
    Language {
        time: u64,
        language: Result<Language, u64>,
    },
    GWBuild {
        time: u64,
        build: u64,
    },
    ShardId {
        time: u64,
        shard: u64,
    },
    Reward(RewardEvent),
    BuffInitial(BuffInitialEvent),
    Position(PositionEvent),
    Velocity(PositionEvent),
    Facing(PositionEvent),
    TeamChange(TeamChangeEvent),
    AttackTarget(AttackTargetEvent),
    Targetable(TargetableEvent),
    MapId {
        time: u64,
        map: u64,
    },
    StackActive(StackActiveEvent),
    StackReset(StackResetEvent),
    Guild(GuildEvent),
    BuffInfo(BuffInfo),
    BuffFormula(BuffFormula),
    SkillInfo(SkillInfo),
    SkillTiming(SkillTiming),
    BreakbarState(BreakbarStateEvent),
    BreakbarPercent(BreakbarPercentEvent),
    Error(ErrorEvent),
    Tag(TagEvent),
    BarrierUpdate(BarrierUpdateEvent),
    StatReset {
        time: u64,
        target: u64,
    },
    Extension {
        sig: u32,
        event: Event,
    },
    ApiDelayed {
        event: Box<EventKind, Global>,
    },
    InstanceStart {
        time: u64,
        start: u64,
    },
    Tickrate {
        time: u64,
        rate: u64,
    },
    Last90BeforeDown(DownContributionEvent),
    EffectOld(EffectOld),
    IdToGUID(EffectGUID),
    LogNPCUpdate(LogEvent),
    ExtensionCombat {
        sig: u32,
        event: Event,
    },
    FractalScale {
        time: u64,
        scale: u64,
    },
    Effect(Effect),
    Activation(ActivationEvent),
    BuffRemove(BuffRemoveEvent),
    BuffApply(BuffApplyEvent),
    BuffDamage(BuffDamageEvent),
    Strike(StrikeEvent),
    Unknown(Event),

}

Possible Event kinds.

Variants

EnterCombat(EnterCombatEvent)

Agent entered combat.

ExitCombat(AgentStatusEvent)

Agent left combat.

ChangeUp(AgentStatusEvent)

Agent is now alive.

ChangeDead(AgentStatusEvent)

Agent is now dead.

ChangeDown(AgentStatusEvent)

Agent is now downed.

Spawn(AgentStatusEvent)

Agent is now in game tracking range.

Despawn(AgentStatusEvent)

Agent is no longer being tracked or out of game tracking range.

HealthUpdate(HealthUpdateEvent)

Agent health change.

LogStart(LogEvent)

Log started.

LogEnd(LogEvent)

Log ended.

WeaponSwap(WeaponSwapEvent)

Agent swapped weapon set.

MaxHealthUpdate(MaxHealthEvent)

Agent maximum health change.

PointOfView(AgentStatusEvent)

Agent is “recording” player.

Language

Fields

time: u64

language: Result<Language, u64>

Game text language.

GWBuild

Fields

time: u64

build: u64

Game build.

ShardId

Fields

time: u64

shard: u64

Sever shard id.

Reward(RewardEvent)

Agent got a reward chest.

BuffInitial(BuffInitialEvent)

Appears once per buff per agent on logging start.

Position(PositionEvent)

Agent position change.

Velocity(PositionEvent)

Agent velocity change.

Facing(PositionEvent)

Agent facing change.

TeamChange(TeamChangeEvent)

Agent team change.

AttackTarget(AttackTargetEvent)

Agent is now an attack target.

Targetable(TargetableEvent)

Agent targetability change.

MapId

Fields

time: u64

map: u64

Map id.

StackActive(StackActiveEvent)

Agent with active buff.

StackReset(StackResetEvent)

Agent with reset buff.

Guild(GuildEvent)

Agent is in guild.

BuffInfo(BuffInfo)

Buff information.

BuffFormula(BuffFormula)

Buff formula.

SkillInfo(SkillInfo)

Skill information.

SkillTiming(SkillTiming)

Skill action.

BreakbarState(BreakbarStateEvent)

Agent breakbar state change.

BreakbarPercent(BreakbarPercentEvent)

Breakbar percentage.

Error(ErrorEvent)

Error.

Tag(TagEvent)

Agent has tag.

BarrierUpdate(BarrierUpdateEvent)

Agent barrier change.

StatReset

Fields

time: u64

target: u64

Arc UI stats reset.

Extension

Fields

sig: u32

event: Event

A custom event created by an extension (addon/plugin).

ApiDelayed

Fields

event: Box<EventKind, Global>

Delayed combat event.

InstanceStart

Fields

time: u64

start: u64

Instance started.

Tickrate

Fields

time: u64

rate: u64

Tick rate.

Last90BeforeDown(DownContributionEvent)

Last 90% before down for downs contribution..

EffectOld(EffectOld)

Effect created or ended.

IdToGUID(EffectGUID)

Id to GUID.

LogNPCUpdate(LogEvent)

Log NPC changed.

ExtensionCombat

Fields

sig: u32

event: Event

A custom combat event created by an extension (addon/plugin).

FractalScale

Fields

time: u64

scale: u64

Fractal scale.

Effect(Effect)

Effect created or ended.

Activation(ActivationEvent)

Activation (cast) event.

BuffRemove(BuffRemoveEvent)

Buff removed.

BuffApply(BuffApplyEvent)

Buff applied.

BuffDamage(BuffDamageEvent)

Buff damage.

Strike(StrikeEvent)

Direct (strike) damage.

Unknown(Event)

Unknown event.