Struct arcdps::evtc::Position

pub struct Position {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

Positional information.

This can be from an Event with StateChange::Position, StateChange::Velocity or StateChange::Facing. It can also occur in StateChange::Effect and StateChange::EffectOld events as effect location or orientation.

Ingame coordinates are interpreted as 1 unit = 1 inch. The z-axis represents vertical height and points down, meaning lower values are a higher location ingame.

Mumble coordinates are given in meters. The y-axis represents vertical height and points up.

Fields
x: f32y: f32z: f32
Implementations

impl Position

pub const INCH_TO_METER: f32 = 0.0253999997f32

Conversion from inch to meter.

pub const fn new(x: f32, y: f32, z: f32) -> Position

Creates new positional information.

pub fn from_mumble(coords: [f32; 3]) -> Position

Creates a position from Mumble coordinates.

pub fn to_array(self) -> [f32; 3]

Converts the position to an array.

pub fn to_tuple(self) -> (f32, f32, f32)

Converts the position to a tuple.

pub fn to_mumble(&self) -> [f32; 3]

Converts the position to Mumble coordinates.

pub fn len(&self) -> f32

Returns the length of the position interpreted as vector.

pub fn mat_mul(&self, matrix: [[f32; 3]; 3]) -> Position

Interprets the position as vector and multiplies it with the given matrix.

pub fn as_rotation_matrix(&self) -> [[f32; 3]; 3]

Interprets the position as rotation angles and converts it to a rotation matrix.

x, y and z are interpreted as angles around each axis in radians.

pub fn rotate(&self, vector: Position) -> Position

Interprets the position as rotation angles and rotates the given vector.