arcdps

Module extras

Unofficial Extras support.

Requires the "extras" feature.

Modules

callbacks

Unofficial extras callback types.

exports

Unofficial Extras exports.

keybinds

Keybind information provided by Unofficial Extras.

language

message

Message information provided by Unofficial Extras.

user

User information provided by Unofficial Extras.

Structs

ExtrasAddonInfo

Information about the Unofficial Extras addon.

ExtrasSubscriberInfo

Information about a subscriber to updates from Unofficial Extras.

ExtrasSubscriberInfoHeader

Subscriber header shared across different versions.

ExtrasVersion

Helper for version checks.

Keybind

A keybind used by the game.

KeybindChange

Keybind change event.

NpcMessage

An NPC chat message.

NpcMessageOwned

NpcMessage with owned String fields.

RawExtrasAddonInfo

SquadMessage

A squad/party chat message.

SquadMessageOwned

SquadMessage with owned String fields.

UserInfo

Information about a player related to the squad.

UserInfoOwned

UserInfo with an owned String name.

Enums

ChannelType

Type of message channel.

Control

A control (player action) used by the game.

Key

A key used by the game.

KeyCode

Custom key codes.

Language

GW2 client language.

Message

A chat message.

MouseCode

Custom mouse codes.

UserRole

Role of a user in the squad.

Functions

strip_account_prefix

Strips the ':' prefix from an account name if present.

subscribe⚠

Subscribes to unofficial extras callbacks after checking for compatibility.

Type Aliases

UserInfoIter

Iterator over changed users.