arcdps::extras

Module exports

Unofficial Extras exports.

Calling an export before Unofficial Extras calls extras_init will cause a panic.

Modules

raw

Raw Unofficial Extras exports.

Functions

get_key

Retrieves the Key for a given game Control from Unofficial Extras. secondary determines whether the primary or secondary bind.

get_keybind

Retrieves the Keybind for a given game Control from Unofficial Extras.

has_get_key

Checks whether the get_key export was found.

has_get_keybind

Checks whether the get_keybind export was found.

version

Retrieves the Unofficial Extras version as string.