Module arcdps::extras::exports::raw

Raw Unofficial Extras exports.

Functions

get_key⚠

Retrieves the RawKey for a given game Control from Unofficial Extras. key_index can be 0 or 1 for primary/secondary keybind respectively.

get_keybind⚠

Retrieves the RawKeybind for a given game Control from Unofficial Extras.

handle⚠

Returns the handle to the Unofficial Extras dll.

Type Aliases

ExportGetKey

Signature of the get_key export.

ExportGetKeybind

Signature of the get_keybind export.