Struct arcdps::extras::RawExtrasAddonInfo

#[repr(C)]
pub struct RawExtrasAddonInfo {
    pub api_version: u32,
    pub max_info_version: u32,
    pub string_version: *const c_char,
    pub self_account_name: *const c_char,
    pub extras_handle: HMODULE,
}

Fields
api_version: u32

Version of the API.

Gets incremented whenever a function signature or behavior changes in a breaking way.

Current version is 2.

max_info_version: u32

Highest known version of the ExtrasSubscriberInfo struct.

Also determines the size of the subscriber info buffer in the init call. The buffer is only guaranteed to have enough space for known ExtrasSubscriberInfo versions.

Current version is 2.

string_version: *const c_char

String version of the Unofficial Extras addon.

Gets changed on every release. The string is valid for the entire lifetime of the Unofficial Extras DLL.

self_account_name: *const c_char

Account name of the logged-in player, including leading ':'.

The string is only valid for the duration of the init call.

extras_handle: HMODULE

The handle to the Unofficial Extras module.

Use this to call the exports of the DLL.

Implementations

impl RawExtrasAddonInfo

pub fn is_compatible(&self) -> bool

Checks compatibility with the Unofficial Extras addon.

pub fn supports_chat_message_callback(&self) -> bool

Whether the Unofficial Extras addon supports the message callback.