arcdps::extras::user

Struct UserInfo

#[repr(C)]
pub struct UserInfo {
    pub join_time: u64,
    pub role: UserRole,
    pub subgroup: u8,
    pub ready_status: bool,
    pub _unused1: u8,
    pub _unused2: u32,
    /* private fields */
}

Information about a player related to the squad.

Strings are available for the duration of the call. If you need it for longer than that, consider converting it to UserInfoOwned.

let owned = user.to_owned();
let owned: UserInfoOwned = user.into();

Fields
join_time: u64

Unix timestamp when the user joined the squad.

0 if time could not be determined.

role: UserRole

Role in squad, or UserRole::None if the user was removed from the squad.

subgroup: u8

Subgroup the user is in.

0 when no subgroup could be found, which is either the first subgroup or no subgroup.

ready_status: bool

Whether this player is ready or not (in a squad ready check).

Remarks

role set to UserRole::SquadLeader and ready_status == true implies that a ready check was just started. Similarly, role set to UserRole::SquadLeader and ready_status == false implies that a ready check either finished or was cancelled. If everyone in the squad had an event sent with ready_status == true then that means that the ready check finished successfully. After which there will be events sent for each user where their ready_status == false.

_unused1: u8

Unused space.

_unused2: u32

Unused space.

Implementations

impl UserInfo

pub fn account_name(&self) -> Option<&str>

Returns the user account name without leading ':'.

pub fn account_name_ptr(&self) -> *const c_char

Returns the raw pointer to the user account name.

pub fn to_owned(&self) -> UserInfoOwned

Converts the UserInfo to the owned version UserInfoOwned.