arcdps

Struct SupportedFields

pub struct SupportedFields {

    pub name: &'static str,
    pub sig: u32,
    pub init: Option<InitFunc>,
    pub release: Option<ReleaseFunc>,
    pub update_url: Option<UpdateUrlFunc>,
    pub raw_wnd_nofilter: Option<RawWndProcCallback>,
    pub raw_imgui: Option<RawImguiCallback>,
    pub raw_options_end: Option<RawOptionsCallback>,
    pub raw_combat: Option<RawCombatCallback>,
    pub raw_wnd_filter: Option<RawWndProcCallback>,
    pub raw_options_windows: Option<RawOptionsWindowsCallback>,
    pub raw_combat_local: Option<RawCombatCallback>,
    pub wnd_nofilter: Option<WndProcCallback>,
    pub combat: Option<CombatCallback>,
    pub imgui: Option<ImguiCallback>,
    pub options_end: Option<OptionsCallback>,
    pub combat_local: Option<CombatCallback>,
    pub wnd_filter: Option<WndProcCallback>,
    pub options_windows: Option<OptionsWindowsCallback>,
    pub raw_extras_init: Option<RawExtrasSubscriberInit>,
    pub raw_extras_squad_update: Option<RawExtrasSquadUpdateCallback>,
    pub raw_extras_language_changed: Option<RawExtrasLanguageChangedCallback>,
    pub raw_extras_keybind_changed: Option<RawExtrasKeybindChangedCallback>,
    pub raw_extras_chat_message: Option<RawExtrasChatMessageCallback>,
    pub extras_init: Option<ExtrasInitFunc>,
    pub extras_squad_update: Option<ExtrasSquadUpdateCallback>,
    pub extras_language_changed: Option<ExtrasLanguageChangedCallback>,
    pub extras_keybind_changed: Option<ExtrasKeybindChangedCallback>,
    pub extras_squad_chat_message: Option<ExtrasSquadChatMessageCallback>,
    pub extras_chat_message: Option<ExtrasChatMessageCallback>,

}

Reference on what fields are currently supported by the export! macro.

This struct is not used anywhere.

Fields
name: &'static str

Name of the plugin.

sig: u32

Unique signature of the plugin.

Pick a random number that is not used by other modules.

init: Option<InitFunc>

Callback for plugin load.

release: Option<ReleaseFunc>

Callback for plugin unload.

update_url: Option<UpdateUrlFunc>

Callback for plugin unload.

raw_wnd_nofilter: Option<RawWndProcCallback>

Raw WndProc callback.

raw_imgui: Option<RawImguiCallback>

Raw ImGui callback.

raw_options_end: Option<RawOptionsCallback>

Raw options callback.

raw_combat: Option<RawCombatCallback>

Raw combat callback.

raw_wnd_filter: Option<RawWndProcCallback>

Raw filtered WndProc callback.

raw_options_windows: Option<RawOptionsWindowsCallback>

Raw options windows callback.

raw_combat_local: Option<RawCombatCallback>

Raw local combat callback.

wnd_nofilter: Option<WndProcCallback>

Callback for key presses.

Returning true will allow ArcDPS and GW2 to receive the key press. First parameter indicates the virtual key code. Second parameter is true if the key was pressed and false when released. Third parameter is true if the key was down before this event occurred, for example by holding it down.

combat: Option<CombatCallback>

Callback for area combat events.

May be called asynchronously, use id to keep track of order. First event id will be 2.

At least one participant will be a party/squad member or minion of, or a buff applied by squad in the case of buff remove. Not all statechanges are present in the realtime API, see StateChange for details.

No event and src.elite == 0 indicates a tracking change. Player was added when src.prof != 0, otherwise removed. When added dst.name contains the account name, dst.id the instance id, dst.prof the Profession, dst.elite the elite Specialization, dst.is_self whether the added player is self (local player), src.team the team and dst.team the subgroup.

No event and src.elite != 0 indicates a target change. src.id will contain the new target.

Note that Arc’s realtime combat API comes with an intentional delay and filtering.

imgui: Option<ImguiCallback>

Callback for standalone UI creation.

Provides an [imgui::Ui] for drawing. The second parameter is true whenever the player is not in character select, loading screens or forced cameras.

options_end: Option<OptionsCallback>

Callback for plugin settings UI creation.

Provides an [imgui::Ui] for drawing.

combat_local: Option<CombatCallback>

Callback for local combat events.

Same as combat but for events from chat log.

wnd_filter: Option<WndProcCallback>

Callback for filtered key presses.

Same as wnd_nofilter but filtered to only notify when modifiers are pressed.

options_windows: Option<OptionsWindowsCallback>

Callback for options windows.

Called for each window checkbox in ArcDPS settings. Last call will always be with None. Does not draw the checkbox if returning true.

raw_extras_init: Option<RawExtrasSubscriberInit>

Raw extras init callback.

Requires the "extras" feature.

raw_extras_squad_update: Option<RawExtrasSquadUpdateCallback>

Raw extras squad update callback.

Requires the "extras" feature.

raw_extras_language_changed: Option<RawExtrasLanguageChangedCallback>

Raw extras language changed callback.

Requires the "extras" feature.

raw_extras_keybind_changed: Option<RawExtrasKeybindChangedCallback>

Raw extras keybind changed callback.

Requires the "extras" feature.

raw_extras_chat_message: Option<RawExtrasChatMessageCallback>

Raw extras chat message callback.

Requires the "extras" feature.

extras_init: Option<ExtrasInitFunc>

Initialization callback for Unofficial Extras.

Can be called before or after ArcDPS init. Receives information about the Unofficial Extras addon and the current player account name as parameters.

Requires the "extras" feature.

extras_squad_update: Option<ExtrasSquadUpdateCallback>

Squad update callback for Unofficial Extras.

Called whenever anything in the squad changes. Only the users that changed are sent. If a user was removed from the squad, their role will be set to UserRole::None.

Requires the "extras" feature.

extras_language_changed: Option<ExtrasLanguageChangedCallback>

Language changed callback for Unofficial Extras.

Called whenever the language is changed, either by changing it in the UI or by pressing the translation key (Right Ctrl by default).

Will be called directly after initialization, with the current language, to get the startup language.

Requires the "extras" feature.

extras_keybind_changed: Option<ExtrasKeybindChangedCallback>

Keybind changed callback for Unofficial Extras.

Called whenever a keybind is changed, either by changing it in the ingame UI or with the presets feature of Unofficial Extras. It is called for every keybind separately.

After initialization this is called for every current keybind that exists. If you want to get a single keybind, at any time you want, call the exported function.

Requires the "extras" feature.

extras_squad_chat_message: Option<ExtrasSquadChatMessageCallback>

Squad chat message callback for Unofficial Extras.

Called whenever a chat message is sent in your party/squad.

Requires the "extras" feature.

extras_chat_message: Option<ExtrasChatMessageCallback>

Chat message callback for Unofficial Extras.

Called on different chat messages.

Requires the "extras" feature.