Crate arcdps

Bindings for ArcDPS plugins.

Usage

Plugins export information for ArcDPS via the export! macro. To see which fields are supported by it, have a look at SupportedFields.

use std::error::Error;
use arcdps::{Agent, CombatEvent, StateChange};

arcdps::export! {
    name: "Example Plugin",
    sig: 0x12345678, // change this to a random number
    init,
    combat: custom_combat_name,
}

fn init() -> Result<(), Box<dyn Error>> {
    // may return an error to indicate load failure
    Ok(())
}

fn custom_combat_name(
    event: Option<CombatEvent>,
    src: Option<Agent>,
    dst: Option<Agent>,
    skill_name: Option<&str>,
    id: u64,
    revision: u64,
) {
    if let Some(event) = event {
        if let StateChange::EnterCombat = event.is_statechange {
            // source agent has entered combat
        }
    }
}

Unofficial Extras

Unofficial Extras support is hidden behind the extras feature flag.

use arcdps::extras::{UserInfoIter, UserRole};

arcdps::export! {
    name: "Example Plugin",
    sig: 123,
    extras_squad_update,
}

fn extras_squad_update(users: UserInfoIter) {
    for user in users {
        if let UserRole::SquadLeader | UserRole::Lieutenant = user.role {
            // user can place markers
        }
    }
}

Re-exports

pub use arcdps_imgui as imgui;

Modules

api

ArcDPS API structs.

callbacks

Callback types.

exports

ArcDPS exports.

extras

Unofficial Extras support.

log

Logging via the log crate.

Macros

export

Creates plugin exports for ArcDPS.

Structs

Agent

Represents an agent in a combat event.

AgentOwned

Agent with an owned String name.

CombatEvent

ArcDPS combat event.

SupportedFields

Reference on what fields are currently supported by the export! macro.

Enums

Activation

Skill activation (cast).

Affinity

Whether the agent is an ally or enemy.

Attribute

Attributes for buff formulas.

BuffCategory

Buff info category.

BuffCycle

Combat buff cycle.

BuffRemove

Combat buff remove.

CustomSkill

ArcDPS custom skill ids.

Language

GW2 client language.

Profession

Player profession.

Specialization

Player specialization.

StateChange

Combat state change.

Strike

Strike types.

Functions

strip_account_prefix

Strips the ':' prefix from an account name if present.