Module arcdps::extras

Unofficial Extras support.

Requires the "extras" feature.

Re-exports

pub use keybinds::Control;

pub use keybinds::Key;

pub use keybinds::KeyCode;

pub use keybinds::KeybindChange;

pub use keybinds::MouseCode;

Modules

callbacks

exports

keybinds

message

Structs

ExtrasAddonInfo

Information about the Unofficial Extras addon.

ExtrasSubscriberInfo

Information about a subscriber to updates from Unofficial Extras.

ExtrasSubscriberInfoHeader

Subscriber header shared across different versions.

RawExtrasAddonInfo

RawUserInfo

UserInfo

Information about a player related to the squad.

UserInfoOwned

Information about a user with owned String.

UserRoleIter

An iterator over the variants of Self

Enums

UserRole

Role of a user in the squad.

Functions

to_user_info_iter⚠

Helper to convert a RawUserInfo pointer and a length to an iterator over UserInfo.

Type Definitions

UserConvert

UserInfoIter

Iterator over changed users.