arcdps::extras::exports

Function get_key

pub fn get_key(control: Control, secondary: bool) -> Key

Retrieves the Key for a given game Control from Unofficial Extras. secondary determines whether the primary or secondary bind.

Returns an empty/default Key if the key is not set or if the functionality is disabled cause of missing patterns.

Examples

use arcdps::extras::{keybinds::Control, exports::get_key};

let primary = get_key(Control::Skills_EliteSkill, false);
let secondary = get_key(Control::Skills_EliteSkill, true);