arcdps

Module log

Logging via the log crate.

Usage

With the "log" feature enabled, ArcDpsLogger is set as logger before your init function is called. By default all messages are logged to ArcDPS’ log window and only warnings & errors are logged to the log file. You can specify "window", "file" or "both"/"all" as log target to control where the messages should be logged.

use log::{error, info};

error!("an error will log to window & file");
error!(target: "window", "window target will only log to window");
error!(target: "file", "file target will only log to file");
info!("below error/warn level will only log to window");
info!(target: "both", "target both/all will log to window & file");

Requires the "log" feature.

Structs

ArcDpsLogger

A logger logging to ArcDPS’ log window and/or file.

FileLogger

A logger logging to ArcDPS’ log file.

WindowLogger

A logger logging to ArcDPS’ log window.